requirements.txt
streamlit
fastapi
uvicorn
anthropic
python-dotenv
requests
flask